requirements.txt
flask
langchain
langchain-groq
langchain-community
chromadb
pypdf
sentence-transformers
python-dotenv
requests
pandas
numpy
